Removed example authenticator, renamed util pkg to auth
