+++ examples/server/java/src/main/java/com/example/helloworld/auth/AuthFilter.java
@@ -1,4 +1,4 @@
-package com.example.helloworld.util;
+package com.example.helloworld.auth;
 
 import java.io.IOException;
 import java.text.ParseException;
