+++ examples/server/java/src/main/java/com/example/helloworld/auth/AuthUtils.java
@@ -1,4 +1,4 @@
-package com.example.helloworld.util;
+package com.example.helloworld.auth;
 
 import java.text.ParseException;
